mcp-client/pyproject.toml
[project]
name = "mcp-client"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "anthropic>=0.49.0",
    "httpx[socks]>=0.28.1",
    "mcp>=1.3.0",
    "openai>=1.68.2",
    "python-dotenv>=1.0.1",
]
